updates to verifyBP, increase accuracy

+++ Training/dRelu.m
@@ -1,3 +1,3 @@
 function [ answer ] = dRelu( x )
-answer=(x>0);%drelu
+answer=double(x>0);%drelu
 end
